Move reordering to library edit page
