+++ src/components/Library/EditLibrary.tsx
@@ -2,10 +2,19 @@ import React, { useEffect, useState } from 'react';
 import { useHistory, useParams } from 'react-router';
 import { toast } from 'react-toastify';
 
-import { Library } from '../../../server/models';
+import { Library, MediaItem } from '../../../server/models';
 import LibraryManager from '../../managers/LibraryManager';
 import { ItemVisibility } from '../../constant/ItemVisibility';
-import { FaCheck, FaTimes, FaTrashAlt } from 'react-icons/fa';
+import { FaCheck, FaLock, FaTimes, FaTrashAlt, FaUnlock } from 'react-icons/fa';
+import {
+  DragDropContext,
+  Droppable,
+  Draggable,
+  DropResult
+} from 'react-beautiful-dnd';
+import { Link } from 'react-router-dom';
+import * as R from 'ramda';
+import MediaManager from '../../managers/MediaManager';
 
 export const EditLibrary = () => {
   const params = useParams<{ library: string }>();
@@ -13,9 +22,13 @@ export const EditLibrary = () => {
 
   const [library, setLibrary] = useState<Library | null>(null);
   const [isDeleting, setIsDeleting] = useState(false);
+  const [media, setMedia] = useState<Array<MediaItem>>([]);
+  const [didReorder, setDidReorder] = useState(false);
+  const [isReordering, setIsReordering] = useState(false);
 
   useEffect(() => {
     LibraryManager.getById(params.library).then(setLibrary);
+    MediaManager.getByLibrary(params.library).then(setMedia);
   }, [params.library]);
 
   const setVisibility = (event: React.ChangeEvent<HTMLSelectElement>) => {
@@ -35,6 +48,22 @@ export const EditLibrary = () => {
       LibraryManager.delete(library._id.toString()).then(() => {
         history.push('/libraries');
       });
+  };
+  const onDragEnd = (result: DropResult) => {
+    if (!result.destination || !library) {
+      return;
+    }
+    setMedia(R.move(result.source.index, result.destination?.index));
+    setDidReorder(true);
+  };
+  const handleReorderClick = () => {
+    if (didReorder) {
+      LibraryManager.update({
+        ...library,
+        items: media.map(R.prop('_id'))
+      } as Library);
+    }
+    setIsReordering(R.not);
   };
 
   return library === null ? null : (
@@ -94,6 +123,60 @@ export const EditLibrary = () => {
               <FaTrashAlt /> Delete
             </button>
           )}
+        </div>
+      </div>
+      <div className="row">
+        <div className="col mt-3 mb-5">
+          <h3 className="h5">
+            Reorder items{' '}
+            <button
+              className="btn btn-outline-warn"
+              onClick={handleReorderClick}
+            >
+              {isReordering ? <FaUnlock /> : <FaLock />}
+            </button>
+          </h3>
+          <DragDropContext onDragEnd={onDragEnd}>
+            <Droppable
+              droppableId="upload-queue"
+              isDropDisabled={!isReordering}
+            >
+              {(provided, snapshot) => (
+                <div {...provided.droppableProps} ref={provided.innerRef}>
+                  <div className="list-group mt-3">
+                    {media.map((mediaItem, index) => (
+                      <Draggable
+                        key={mediaItem._id.toString()}
+                        draggableId={mediaItem._id.toString()}
+                        index={index}
+                      >
+                        {(provided, snapshot) => (
+                          <div
+                            ref={provided.innerRef}
+                            {...provided.draggableProps}
+                            {...provided.dragHandleProps}
+                          >
+                            <div
+                              className="list-group-item d-flex justify-content-between"
+                              key={mediaItem._id.toString()}
+                            >
+                              <Link
+                                to={`/library/${library._id.toString()}/${mediaItem._id.toString()}`}
+                                key={mediaItem._id.toString()}
+                              >
+                                {mediaItem.name}{' '}
+                              </Link>
+                            </div>
+                          </div>
+                        )}
+                      </Draggable>
+                    ))}
+                    {provided.placeholder}
+                  </div>
+                </div>
+              )}
+            </Droppable>
+          </DragDropContext>
         </div>
       </div>
     </div>
